Split SpriteAnimator into several classes
Factor out Anime class
Factor out Sprite class
Factor out Database to class
Factor out GUI construction to class
Factor out GUI Helper functions to class
Factor out `.spr` Manipulation functions to class
