+++ GuiHelpers.java
@@ -1,0 +1,75 @@
+package SpriteAnimator;
+
+public class GuiHelpers {
+	/*
+	 * GUI related functions
+	 */
+	/**
+	 * gives file extension name from a string
+	 * @param s - test case
+	 * @return extension type
+	 */
+	public static String getFileType(String s) {
+		String ret = s.substring(s.lastIndexOf(".") + 1);
+		return ret;
+	}
+
+	/**
+	 * Test a file against multiple extensions.
+	 * The way <b>getFileType</b> works should allow
+	 * both full paths and lone file types to work.
+	 * 
+	 * @param s - file name or extension
+	 * @param type - list of all extensions to test against
+	 * @return <tt>true</tt> if any extension is matched
+	 */
+	public static boolean testFileType(String s, String[] type) {
+		boolean ret = false;
+		String filesType = getFileType(s);
+		for (String t : type) {
+			if (filesType.equalsIgnoreCase(t)) {
+				ret = true;
+				break;
+			}
+		}
+		return ret;
+	}
+
+	/**
+	 * Test a file against a single extension.
+	 * 
+	 * @param s - file name or extension
+	 * @param type - extension
+	 * @return <tt>true</tt> if extension is matched
+	 */
+	public boolean testFileType(String s, String type) {
+		return testFileType(s, new String[] { type });
+	}
+	
+	/**
+	 * Join array of strings together with a delimiter.
+	 * @param s - array of strings
+	 * @param c - delimiter
+	 * @return A single <tt>String</tt>.
+	 */
+	public String join(String[] s, String c) {
+		String ret = "";
+		for (int i = 0; i < s.length; i++) {
+			ret += s[i];
+			if (i != s.length-1) {
+				ret += c;
+			}
+		}
+		return ret;
+	}
+
+	/**
+	 * 
+	 * @param b
+	 * @return
+	 */
+	public int unsignByte(byte b) {
+		int ret = (b + 256) % 256;
+		return ret;
+	}
+}
