+++ SprManipulator.java
@@ -1,0 +1,188 @@
+package SpriteAnimator;
+
+import java.awt.image.BufferedImage;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
+import java.io.IOException;
+
+public class SprManipulator {
+	static final GuiHelpers GH = new GuiHelpers();
+
+	static final int SPRITESIZE = 896 * 32; // invariable lengths
+	static final int RASTERSIZE = 128 * 448 * 4;
+	
+	// format of snes 4bpp {row (r), bit plane (b)}
+	// bit plane 0 indexed such that 1011 corresponds to 0123
+	static final int BPPI[][] = {
+			{0,0},{0,1},{1,0},{1,1},{2,0},{2,1},{3,0},{3,1},
+			{4,0},{4,1},{5,0},{5,1},{6,0},{6,1},{7,0},{7,1},
+			{0,2},{0,3},{1,2},{1,3},{2,2},{2,3},{3,2},{3,3},
+			{4,2},{4,3},{5,2},{5,3},{6,2},{6,3},{7,2},{7,3}
+	};
+
+	/**
+	 * Reads a sprite file
+	 * @throws IOException
+	 */
+	public byte[] readSprite(String path) throws IOException {
+		File file = new File(path);
+		byte[] ret = new byte[(int) file.length()];
+		FileInputStream s;
+		try {
+			s = new FileInputStream(file);
+		} catch (FileNotFoundException e) {
+			throw e;
+		}
+		try {
+			s.read(ret);
+			s.close();
+		} catch (IOException e) {
+			throw e;
+		}
+
+		return ret;
+	}
+
+	/**
+	 * Takes a sprite and turns it into 896 blocks of 8x8 pixels
+	 * @param sprite
+	 */
+	public byte[][][] sprTo8x8(byte[] sprite) {
+		byte[][][] ret = new byte[896][8][8];
+
+		// current block we're working on, each sized 32
+		// start at -1 since we're incrementing at 0mod32
+		int b = -1;
+		// locate where in interlacing map we're reading from
+		int g;
+		for (int i = 0; i < SPRITESIZE; i++) {
+			// find interlacing index
+			g = i%32;
+			// increment at 0th index
+			if (g == 0) {
+				b++;
+			}
+			// row to look at
+			int r = BPPI[g][0];
+			// bit plane of byte
+			int p = BPPI[g][1];
+
+			// byte to unravel
+			byte q = sprite[i];
+
+			// run through the byte
+			for (int c = 0; c < 8; c++) {
+				// AND with 1 shifted to the correct plane
+				boolean bitOn = (q & (1 << (7-c))) != 0;
+				// if true, OR with that plane in index map
+				if (bitOn) {
+					ret[b][r][c] |= (1 << (p));
+				}
+			}
+		}
+		return ret;
+	}
+
+	/**
+	 * Splits a palette into RGB arrays.
+	 * Only uses the first 16 colors.
+	 * Automatically makes first index black.
+	 */
+	public byte[][] getPal(byte[] sprite) {
+		byte[][] ret = new byte[16][3];
+		for (int i = 1; i < 16; i++) {
+			short color = 0;
+			int pos = SPRITESIZE + (i * 2) - 2;
+			color = (short) GH.unsignByte(sprite[pos+1]);
+			color <<= 8;
+			color |= (short) GH.unsignByte(sprite[pos]);
+			
+			ret[i][0] = (byte) (((color >> 0) & 0x1F) << 3);
+			ret[i][1] = (byte) (((color >> 5) & 0x1F) << 3);
+			ret[i][2] = (byte) (((color >> 10) & 0x1F) << 3);
+		}
+
+		// make black;
+		// separate operation just in case I don't wanna change pal's values
+		ret[0][0] = 0;
+		ret[0][1] = 0;
+		ret[0][2] = 0;
+
+		return ret;
+	}
+
+	/**
+	 * Turn index map in 8x8 format into an array of ABGR values
+	 */
+	public byte[] makeRaster(byte[][][] ebe, byte[][] palette) {
+		byte[] ret = new byte[RASTERSIZE];
+		int largeCol = 0;
+		int intRow = 0;
+		int intCol = 0;
+		int index = 0;
+		byte[] color;
+		// read image
+		for (int i = 0; i < RASTERSIZE / 4; i++) {
+			// get pixel color index
+			byte coli = ebe[index][intRow][intCol];
+			// get palette color
+			color = palette[coli];
+			// index 0 = trans
+			if (coli == 0) {
+				ret[i*4] = 0;
+			}
+			else {
+				ret[i*4] = (byte) 255;
+			}
+
+			// BGR
+			ret[i*4+1] = color[2];
+			ret[i*4+2] = color[1];
+			ret[i*4+3] = color[0];
+
+			// count up square by square
+			// at 8, reset the "Interior column" which we use to locate the pixel in 8x8
+			// increments the "Large column", which is the index of the 8x8 sprite on the sheet
+			// at 16, reset the index and move to the next row
+			// (so we can wrap around back to our old 8x8)
+			// after 8 rows, undo the index reset, and move on to the next super row
+			intCol++;
+			if (intCol == 8) {
+				index++;
+				largeCol++;
+				intCol = 0;
+				if (largeCol == 16) {
+					index -= 16;
+					largeCol = 0;
+					intRow++;
+					if (intRow == 8) {
+						index += 16;
+						intRow = 0;
+					}
+				}
+			}
+		}
+		return ret;
+	}
+
+	/**
+	 * Turns a 4 byte raster {A,B,G,R} into an integer array and sets the image.
+	 * @param raster
+	 * @return
+	 */
+	public BufferedImage makeSheet(byte[] raster) {
+		BufferedImage image = new BufferedImage(128, 448, BufferedImage.TYPE_4BYTE_ABGR_PRE);
+		int[] rgb = new int[128 * 448];
+		for (int i = 0, j = 0; i < rgb.length; i++) {
+			int a = raster[j++] & 0xff;
+			int b = raster[j++] & 0xff;
+			int g = raster[j++] & 0xff;
+			int r = raster[j++] & 0xff;
+			rgb[i] = (a << 24) | (r << 16) | (g << 8) | b;
+		}
+		image.setRGB(0, 0, 128, 448, rgb, 0, 128);
+		
+		return image;
+	}
+}
